+++ sdmx-dl-api/src/main/java/sdmxdl/web/MonitorReport.java
@@ -2,8 +2,6 @@ package sdmxdl.web;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-import java.net.URL;
-
 @lombok.Value
 @lombok.Builder(toBuilder = true)
 public class MonitorReport {
@@ -22,8 +20,4 @@ public class MonitorReport {
     @Nullable
     @lombok.Builder.Default
     Long averageResponseTime = null;
-
-    @Nullable
-    @lombok.Builder.Default
-    URL webReport = null;
 }