+++ sdmx-dl-format-kryo/src/main/java/sdmxdl/format/kryo/KryoFileFormat.java
@@ -485,7 +485,6 @@ public final class KryoFileFormat<T> implements FileParser<T>, FileFormatter<T> 
             kryo.writeObject(output, t.getStatus());
             kryo.writeObjectOrNull(output, t.getUptimeRatio(), Double.class);
             kryo.writeObjectOrNull(output, t.getAverageResponseTime(), Long.class);
-            kryo.writeObjectOrNull(output, t.getWebReport(), URL.class);
         }
 
         @Override
@@ -496,7 +495,6 @@ public final class KryoFileFormat<T> implements FileParser<T>, FileFormatter<T> 
                     .status(kryo.readObject(input, MonitorStatus.class))
                     .uptimeRatio(kryo.readObjectOrNull(input, Double.class))
                     .averageResponseTime(kryo.readObjectOrNull(input, Long.class))
-                    .webReport(kryo.readObjectOrNull(input, URL.class))
                     .build();
         }
     }