+++ sdmx-dl-format-xml/src/main/java/sdmxdl/format/xml/XmlWebSource.java
@@ -61,6 +61,7 @@ public final class XmlWebSource {
     private static final String KEY_ATTR = "key";
     private static final String VALUE_ATTR = "value";
     private static final String MONITOR_TAG = "monitor";
+    private static final String MONITOR_WEBSITE_TAG = "monitorWebsite";
 
     private static List<SdmxWebSource> parseXml(XMLStreamReader reader) throws XMLStreamException {
         List<SdmxWebSource> result = new ArrayList<>();
@@ -100,6 +101,9 @@ public final class XmlWebSource {
                         case MONITOR_TAG:
                             item.monitorOf(reader.getElementText());
                             break;
+                        case MONITOR_WEBSITE_TAG:
+                            item.monitorWebsiteOf(reader.getElementText());
+                            break;
                     }
                     break;
                 case XMLStreamReader.END_ELEMENT:
@@ -134,6 +138,7 @@ public final class XmlWebSource {
             }
             writeTextElement(writer, WEBSITE_TAG, source.getWebsite());
             writeTextElement(writer, MONITOR_TAG, source.getMonitor());
+            writeTextElement(writer, MONITOR_WEBSITE_TAG, source.getMonitorWebsite());
             writer.writeEndElement();
         }
         writer.writeEndElement();