+++ sdmx-dl-provider-connectors/src/main/java/internal/sdmxdl/provider/connectors/drivers/EurostatDriver.java
@@ -51,6 +51,7 @@ public final class EurostatDriver implements WebDriver {
                     .endpointOf("https://ec.europa.eu/eurostat/SDMX/diss-web/rest")
                     .websiteOf("https://ec.europa.eu/eurostat/data/database")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/ESTAT")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/estat")
                     .build())
             .build();
 }