+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/provider/ri/web/drivers/Sdmx21Driver2.java
@@ -60,6 +60,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://explore.data.abs.gov.au")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/ABS")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/abs")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -70,6 +71,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://stats.bis.org/statx/toc/LBS.html")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/BIS")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/bis")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -79,6 +81,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .endpointOf("https://nsiws-stable-camstat-live.officialstatistics.org/rest")
                     .websiteOf("http://camstat.nis.gov.kh/?locale=en&start=0")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/CAMSTAT")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/camstat")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -114,6 +117,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://sdw.ecb.europa.eu")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/ECB")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/ecb")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -123,6 +127,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .endpointOf("https://api-dataexplorer.unescap.org/rest/")
                     .websiteOf("https://dataexplorer.unescap.org/")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/ESCAP")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/escap")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -136,6 +141,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://ilostat.ilo.org/data/")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/ILO")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/ilo")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -148,6 +154,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://data.imf.org")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/IMF_SDMX_CENTRAL")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/imf-sdmx-central")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -161,6 +168,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(TRAILING_SLASH_REQUIRED_PROPERTY, true)
                     .websiteOf("https://sdmx.snieg.mx")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/INEGI")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/inegi")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -173,6 +181,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://www.istat.it/en/analysis-and-products")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/ISTAT")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/istat")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -183,6 +192,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://www.norges-bank.no/en/topics/Statistics/")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/NB")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/nb")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -193,6 +203,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://registry.sdmx.org/overview.html")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/SGR")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/sgr")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -205,6 +216,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://datasimel.mtps.gob.sv/")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/SIMEL")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/simel")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -215,6 +227,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://stats.pacificdata.org/?locale=en")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/SPC")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/spc")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -227,6 +240,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://lustat.statec.lu")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/STATEC")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/statec")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -237,6 +251,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
                     .websiteOf("https://data.un.org/SdmxBrowser/start")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/UNDATA")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/undata")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -251,6 +266,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(TRAILING_SLASH_REQUIRED_PROPERTY, true)
                     .websiteOf("https://data.worldbank.org")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/WB")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/wb")
                     .build())
             .source(SdmxWebSource
                     .builder()
@@ -263,6 +279,7 @@ public final class Sdmx21Driver2 implements WebDriver {
                     .propertyOf(TRAILING_SLASH_REQUIRED_PROPERTY, true)
                     .websiteOf("https://wits.worldbank.org")
                     .monitorOf("upptime:/nbbrd/sdmx-upptime/WITS")
+                    .monitorWebsiteOf("https://nbbrd.github.io/sdmx-upptime/history/wits")
                     .build())
             .build();